modify business profiles logic to work with "unclaimed" businesses not having content or organization (#627)

+++ app/models/business_profile.rb
@@ -23,7 +23,9 @@ class BusinessProfile < ActiveRecord::Base
   validates_associated :content
 
   after_destroy do
-    organization.destroy if organization.present? and organization.contents.count == 0
+    if content.present?
+      organization.destroy if organization.present? and organization.contents.count == 0
+    end
   end
 
   belongs_to :business_location, dependent: :destroy
